Update tst_bench_qhash hash functions to use size_t
Pre-requisite for a fix for qHash. The Qt50String inherits from QString
and becomes ambiguous once it no longer goes through a catch-all
template function because qHash(QString, size_t) has a better match for
the second argument.

Pick-to: 6.2 6.4

+++ tests/benchmarks/corelib/tools/qhash/tst_bench_qhash.h
@@ -10,7 +10,7 @@ struct Qt4String : QString
 };
 
 QT_BEGIN_NAMESPACE
-uint qHash(const Qt4String &);
+size_t qHash(const Qt4String &);
 QT_END_NAMESPACE
 
 struct Qt50String : QString
@@ -20,7 +20,7 @@ struct Qt50String : QString
 };
 
 QT_BEGIN_NAMESPACE
-uint qHash(const Qt50String &, uint seed = 0);
+size_t qHash(const Qt50String &, size_t seed = 0);
 QT_END_NAMESPACE
 
 
@@ -31,6 +31,6 @@ struct JavaString : QString
 };
 
 QT_BEGIN_NAMESPACE
-uint qHash(const JavaString &);
+size_t qHash(const JavaString &);
 QT_END_NAMESPACE
